Updated release and branch release scripts.

+++ bacnet-stack-0-6-0/release-branch.sh
@@ -8,6 +8,7 @@ FRS_URL=skarg,bacnet@frs.sourceforge.net:/home/frs/project/b/ba/bacnet/bacnet-st
 if [ -z "$1" ]
 then
   echo "Usage: `basename $0` 0.0.0 0.0.1"
+  echo "Use dotted version as [branch-revision] [tag-revision]."
   echo "Creates the ChangeLog."
   echo "Creates the release files."
   echo "Tags this branch release in subversion."
@@ -23,14 +24,18 @@ echo "Creating the ${TAGGED_VERSION_DOTTED} release files for $(BRANCH_VERSION_D
 
 CHANGELOG=ChangeLog-${TAGGED_VERSION_DOTTED}
 echo "Creating the ${PROJECT} change log ${CHANGELOG}"
+if [ -e "${CHANGELOG}" ]
+then
 rm ${CHANGELOG}
+fi
 svn update
 svn log --xml --verbose | xsltproc svn2cl.xsl - > ${CHANGELOG}
-if [ -z "${CHANGELOG}" ]
+if [ -e "${CHANGELOG}" ]
 then
-echo "Failed to create ${CHANGELOG}"
+  echo "${CHANGELOG} created."
 else
-echo "${CHANGELOG} created."
+  echo "Failed to create ${CHANGELOG}"
+  exit 1
 fi
 
 BRANCH_NAME=${SVN_MODULE}-${BRANCH_VERSION_DASHED}
@@ -41,44 +46,65 @@ SVN_BASE_URL=https://${PROJECT}.svn.sourceforge.net/svnroot/${PROJECT}
 SVN_BRANCH_NAME=${SVN_BASE_URL}/branches/releases/${BRANCH_NAME}
 SVN_TAGGED_NAME=${SVN_BASE_URL}/tags/${TAGGED_NAME}
 echo "Setting a tag on the ${SVN_MODULE} module called ${TAGGED_NAME}"
-svn copy ${SVN_BRANCH_NAME} ${SVN_TAGGED_NAME} -m "Created version ${ARCHIVE_NAME}"
+svn copy ${SVN_BRANCH_NAME} ${SVN_TAGGED_NAME} -m "tagged" > /dev/null
 echo "done."
 
+if [ -d "${ARCHIVE_NAME}" ]
+then
+  echo "removing old ${ARCHIVE_NAME}..."
+  rm -rf ${ARCHIVE_NAME}
+  echo "done."
+fi
+
 echo "Getting a clean version out of subversion for Linux gzip"
-svn export ${SVN_TAGGED_NAME} ${ARCHIVE_NAME}
+svn export ${SVN_TAGGED_NAME} ${ARCHIVE_NAME} > /dev/null
 echo "done."
 
 GZIP_FILENAME=${ARCHIVE_NAME}.tgz
 echo "tar and gzip the clean directory"
-tar -cvvzf ${GZIP_FILENAME} ${ARCHIVE_NAME}/
-echo "done."
-
-if [ -z "${GZIP_FILENAME}" ]
+if [ -e "${GZIP_FILENAME}" ]
 then
-echo "Failed to create ${GZIP_FILENAME}"
+  echo "removing old ${GZIP_FILENAME}..."
+  rm ${GZIP_FILENAME}
+  echo "done."
+fi
+tar -cvvzf ${GZIP_FILENAME} ${ARCHIVE_NAME}/ > /dev/null
+echo "done."
+if [ -e "${GZIP_FILENAME}" ]
+then
+  echo "${GZIP_FILENAME} created."
 else
-echo "${GZIP_FILENAME} created."
+  echo "Failed to create ${GZIP_FILENAME}"
+  exit 1
 fi
 
-echo "Removing the directory exported for Linux."
-rm -rf ${ARCHIVE_NAME}
-
+if [ -d "${ARCHIVE_NAME}" ]
+then
+  echo "removing old ${ARCHIVE_NAME}..."
+  rm -rf ${ARCHIVE_NAME}
+  echo "done."
+fi
 echo "Getting another clean version out of subversion for Windows zip"
-svn export --native-eol CRLF ${SVN_TAGGED_NAME} ${ARCHIVE_NAME}
+svn export --native-eol CRLF ${SVN_TAGGED_NAME} ${ARCHIVE_NAME} > /dev/null
 ZIP_FILENAME=${ARCHIVE_NAME}.zip
 echo "done."
 echo "Zipping the directory exported for Windows."
-zip -r ${ZIP_FILENAME} ${ARCHIVE_NAME}
-
-if [ -z "${ZIP_FILENAME}" ]
+zip -r ${ZIP_FILENAME} ${ARCHIVE_NAME} > /dev/null
+if [ -e "${ZIP_FILENAME}" ]
 then
-echo "Failed to create ${ZIP_FILENAME}"
+  echo "${ZIP_FILENAME} created."
 else
-echo "${ZIP_FILENAME} created."
+  echo "Failed to create ${ZIP_FILENAME}"
+  exit 1
 fi
 
-echo "Removing the directory exported for Windows."
-rm -rf ${ARCHIVE_NAME}
+# remove SVN files
+if [ -d "${ARCHIVE_NAME}" ]
+then
+  echo "removing ${ARCHIVE_NAME}..."
+  rm -rf ${ARCHIVE_NAME}
+  echo "done."
+fi
 
 echo "Creating ${ARCHIVE_NAME}"
 mkdir ${ARCHIVE_NAME}
