+++ bacnet-stack-0-6-0/release.sh
@@ -21,14 +21,18 @@ echo "Creating the release files for version ${DOTTED_VERSION}"
 
 CHANGELOG=ChangeLog-${DOTTED_VERSION}
 echo "Creating the ${PROJECT} change log ${CHANGELOG}"
+if [ -e "${CHANGELOG}" ]
+then
 rm ${CHANGELOG}
+fi
 svn update
 svn log --xml --verbose | xsltproc svn2cl.xsl - > ${CHANGELOG}
-if [ -z "${CHANGELOG}" ]
+if [ -e "${CHANGELOG}" ]
 then
-echo "Failed to create ${CHANGELOG}"
+  echo "${CHANGELOG} created."
 else
-echo "${CHANGELOG} created."
+  echo "Failed to create ${CHANGELOG}"
+  exit 1
 fi
 
 ARCHIVE_NAME=${SVN_MODULE}-${DOTTED_VERSION}
@@ -38,42 +42,66 @@ SVN_BASE_URL=https://${PROJECT}.svn.sourceforge.net/svnroot/${PROJECT}
 SVN_TRUNK_NAME=${SVN_BASE_URL}/trunk/${SVN_MODULE}
 SVN_TAGGED_NAME=${SVN_BASE_URL}/tags/${TAGGED_NAME}
 echo "Setting a tag on the ${SVN_MODULE} module called ${TAGGED_NAME}"
-svn copy ${SVN_TRUNK_NAME} ${SVN_TAGGED_NAME} -m "Created version ${ARCHIVE_NAME}"
+svn copy ${SVN_TRUNK_NAME} ${SVN_TAGGED_NAME} -m "tagged" > /dev/null
 echo "done."
 
+if [ -d "${ARCHIVE_NAME}" ]
+then
+  echo "removing old ${ARCHIVE_NAME}..."
+rm -rf ${ARCHIVE_NAME}
+  echo "done."
+fi
+
 echo "Getting a clean version out of subversion for Linux gzip"
-svn export ${SVN_TAGGED_NAME} ${ARCHIVE_NAME}
+svn export ${SVN_TAGGED_NAME} ${ARCHIVE_NAME} > /dev/null
 echo "done."
 
 GZIP_FILENAME=${ARCHIVE_NAME}.tgz
 echo "tar and gzip the clean directory"
-tar -cvvzf ${GZIP_FILENAME} ${ARCHIVE_NAME}/
-echo "done."
-
-if [ -z "${GZIP_FILENAME}" ]
+if [ -e "${GZIP_FILENAME}" ]
 then
-echo "Failed to create ${GZIP_FILENAME}"
+  echo "removing old ${GZIP_FILENAME}..."
+  rm ${GZIP_FILENAME}
+  echo "done."
+fi
+tar -cvvzf ${GZIP_FILENAME} ${ARCHIVE_NAME}/ > /dev/null
+echo "done."
+if [ -e "${GZIP_FILENAME}" ]
+then
+  echo "${GZIP_FILENAME} created."
 else
-echo "${GZIP_FILENAME} created."
+  echo "Failed to create ${GZIP_FILENAME}"
+  exit 1
 fi
 
-rm -rf ${ARCHIVE_NAME}
-
+if [ -d "${ARCHIVE_NAME}" ]
+then
+  echo "removing old ${ARCHIVE_NAME}..."
+  rm -rf ${ARCHIVE_NAME}
+  echo "done."
+fi
 echo "Getting another clean version out of subversion for Windows zip"
-svn export --native-eol CRLF ${SVN_TAGGED_NAME} ${ARCHIVE_NAME}
+svn export --native-eol CRLF ${SVN_TAGGED_NAME} ${ARCHIVE_NAME} > /dev/null
 ZIP_FILENAME=${ARCHIVE_NAME}.zip
 echo "done."
 echo "Zipping the directory exported for Windows."
-zip -r ${ZIP_FILENAME} ${ARCHIVE_NAME}
+zip -r ${ZIP_FILENAME} ${ARCHIVE_NAME} > /dev/null
 
-if [ -z "${ZIP_FILENAME}" ]
+if [ -e "${ZIP_FILENAME}" ]
 then
-echo "Failed to create ${ZIP_FILENAME}"
+  echo "${ZIP_FILENAME} created."
 else
-echo "${ZIP_FILENAME} created."
+  echo "Failed to create ${ZIP_FILENAME}"
+  exit 1
 fi
 
-rm -rf ${ARCHIVE_NAME}
+# remove SVN files
+if [ -d "${ARCHIVE_NAME}" ]
+then
+  echo "removing ${ARCHIVE_NAME}..."
+  rm -rf ${ARCHIVE_NAME}
+  echo "done."
+fi
 
 echo "Creating ${ARCHIVE_NAME}"
 mkdir ${ARCHIVE_NAME}
